core_engine/requirements.txt
opencv-python
openai-whisper
moviepy
torch
pytest
numpy
